Adding debian/ folder needed to build the package.
Created environment (debian/) to build package using dpkg-buildpackag
tool. This is needed to move from the automation process that was
used by intel and to build only packages we care about, e.g libpmem.

Even though it is possbile to build this package, it still has several
"warning" and "error" messages that needed to be fixed such as:
    - E: nvml source: source-is-missing src/tools/pmempool/pmempool.static-debug
    - W: nvml source: configure-generated-file-in-source src/nondebug/jemalloc/config.status
    - W: nvml source: maintainer-script-lacks-debhelper-token debian/libpmem.postinst
    - W: nvml source: ancient-standards-version 3.9.4 (current is 3.9.6)

Packages that are built:
    - libpmem
    - libpmem-dev