Fixing lintian errors and warning messages.
Previously we successfully packaged NVML code, but Lintian had several
warnings and errors that needed to be fixed. In this commit most of the
errors were physically fixed, e.g. appropriate files were slightly modified.
Error messages complaining "source-is-missing" were overridden.

Justin Vreeland:

Since this is packaged as a debian-native package there aren't revision
numbers.  We find this acceptable for ease of packaging.